Extract Qt atomic operation wrappers into separate header file
