+++ src/engine/cachingreader/cachingreaderworker.cpp
@@ -8,7 +8,7 @@
 #include "moc_cachingreaderworker.cpp"
 #include "sources/soundsourceproxy.h"
 #include "track/track.h"
-#include "util/compatibility.h"
+#include "util/compatibility/qatomic.h"
 #include "util/compatibility/qmutex.h"
 #include "util/event.h"
 #include "util/logger.h"
