+++ src/engine/enginebuffer.cpp
@@ -30,7 +30,7 @@
 #include "track/keyutils.h"
 #include "track/track.h"
 #include "util/assert.h"
-#include "util/compatibility.h"
+#include "util/compatibility/qatomic.h"
 #include "util/defs.h"
 #include "util/logger.h"
 #include "util/sample.h"
