+++ src/engine/sidechain/shoutconnection.cpp
@@ -28,6 +28,7 @@
 #include "preferences/usersettings.h"
 #include "recording/defs_recording.h"
 #include "track/track.h"
+#include "util/compatibility/qatomic.h"
 #include "util/logger.h"
 
 namespace {
