+++ src/mixer/playermanager.cpp
@@ -19,6 +19,7 @@
 #include "soundio/soundmanager.h"
 #include "track/track.h"
 #include "util/assert.h"
+#include "util/compatibility/qatomic.h"
 #include "util/defs.h"
 #include "util/logger.h"
 #include "util/sleepableqthread.h"
