+++ src/preferences/dialog/dlgprefinterface.cpp
@@ -17,6 +17,7 @@
 #include "skin/legacy/legacyskinparser.h"
 #include "skin/skin.h"
 #include "skin/skinloader.h"
+#include "util/compatibility.h"
 #include "util/screensaver.h"
 #include "util/screensavermanager.h"
 #include "util/widgethelper.h"
