+++ src/soundio/soundmanager.cpp
@@ -19,7 +19,7 @@
 #include "soundio/sounddeviceportaudio.h"
 #include "soundio/soundmanagerutil.h"
 #include "util/cmdlineargs.h"
-#include "util/compatibility.h"
+#include "util/compatibility/qatomic.h"
 #include "util/defs.h"
 #include "util/sample.h"
 #include "util/sleep.h"
