+++ src/util/compatibility.h
@@ -1,5 +1,7 @@
 #pragma once
 
+// TODO: Split this file into separate header files in subdirectory util/compatibility
+
 #include <QCoreApplication>
 #include <QGuiApplication>
 #include <QScreen>
@@ -35,88 +37,4 @@ QString uuidToNullableStringWithoutBraces(const QUuid& uuid) {
     } else {
         return uuidToStringWithoutBraces(uuid);
     }
-}
-
-template <typename T>
-inline T atomicLoadAcquire(const QAtomicInteger<T>& atomicInt) {
-    // TODO: QBasicAtomicInteger<T>::load() is deprecated and should be
-    // replaced with QBasicAtomicInteger<T>::loadRelaxed() However, the
-    // proposed alternative has just been introduced in Qt 5.14. Until the
-    // minimum required Qt version of Mixxx is increased, we need a version
-    // check here
-#if QT_VERSION >= QT_VERSION_CHECK(5, 14, 0)
-    return atomicInt.loadAcquire();
-#else
-    return atomicInt.load();
-#endif
-}
-
-template <typename T>
-inline T* atomicLoadAcquire(const QAtomicPointer<T>& atomicPtr) {
-    // TODO: QBasicAtomicPointer<T>::load() is deprecated and should be
-    // replaced with QBasicAtomicPointer<T>::loadRelaxed() However, the
-    // proposed alternative has just been introduced in Qt 5.14. Until the
-    // minimum required Qt version of Mixxx is increased, we need a version
-    // check here
-#if QT_VERSION >= QT_VERSION_CHECK(5, 14, 0)
-    return atomicPtr.loadAcquire();
-#else
-    return atomicPtr.load();
-#endif
-}
-
-template <typename T>
-inline T atomicLoadRelaxed(const QAtomicInteger<T>& atomicInt) {
-    // TODO: QBasicAtomicInteger<T>::load() is deprecated and should be
-    // replaced with QBasicAtomicInteger<T>::loadRelaxed() However, the
-    // proposed alternative has just been introduced in Qt 5.14. Until the
-    // minimum required Qt version of Mixxx is increased, we need a version
-    // check here
-#if QT_VERSION >= QT_VERSION_CHECK(5, 14, 0)
-    return atomicInt.loadRelaxed();
-#else
-    return atomicInt.load();
-#endif
-}
-
-template <typename T>
-inline T* atomicLoadRelaxed(const QAtomicPointer<T>& atomicPtr) {
-    // TODO: QBasicAtomicPointer<T>::load() is deprecated and should be
-    // replaced with QBasicAtomicPointer<T>::loadRelaxed() However, the
-    // proposed alternative has just been introduced in Qt 5.14. Until the
-    // minimum required Qt version of Mixxx is increased, we need a version
-    // check here
-#if QT_VERSION >= QT_VERSION_CHECK(5, 14, 0)
-    return atomicPtr.loadRelaxed();
-#else
-    return atomicPtr.load();
-#endif
-}
-
-template <typename T>
-inline void atomicStoreRelaxed(QAtomicInteger<T>& atomicInt, T newValue) {
-    // TODO: QBasicAtomicInteger<T>::store(T newValue) is deprecated and should
-    // be replaced with QBasicAtomicInteger<T>::storeRelaxed(T newValue)
-    // However, the proposed alternative has just been introduced in Qt 5.14.
-    // Until the minimum required Qt version of Mixxx is increased, we need a
-    // version check here
-#if QT_VERSION >= QT_VERSION_CHECK(5, 14, 0)
-    atomicInt.storeRelaxed(newValue);
-#else
-    atomicInt.store(newValue);
-#endif
-}
-
-template <typename T>
-inline void atomicStoreRelaxed(QAtomicPointer<T>& atomicPtr, T* newValue) {
-    // TODO: QBasicAtomicPointer<T>::store(T* newValue) is deprecated and
-    // should be replaced with QBasicAtomicPointer<T>::storeRelaxed(T*
-    // newValue) However, the proposed alternative has just been introduced in
-    // Qt 5.14. Until the minimum required Qt version of Mixxx is increased, we
-    // need a version check here
-#if QT_VERSION >= QT_VERSION_CHECK(5, 14, 0)
-    atomicPtr.storeRelaxed(newValue);
-#else
-    atomicPtr.store(newValue);
-#endif
 }
